nioldos: fix grub-legacy boot
Binaries exported by BS contain debug symbols and OE by default is
configured to strip all bins it ships in non dbg packages. This causes
errors with grub-legacy which can't boot (error "no symbol tables found").

It is safer to let BS binaries behave just as BS exported and not modify
them because this can cause undefined behaviour. The size of these debug
symbols is negligible compared to the whole size of the nioldos image ipk.

If we really want to strip these files then we should do it in BS and OE
with this change will simply use what binaries BS produces/exports.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>