nioldos: refactor to use the NXG restore-mode logic
nioldos is the only migration path which doesn't use the already
existing restore-mode cpio infrastructure, in contrast with fw/bw arm
and fw x64 migration, it hijacks a runmode image boot & hacks up
the BS-built/export fetched runmode image to provision a safemode.

There are a lot of problems with this approach, the most glaring of
which is the code ugliness. The runmode image is bloated (eg. it has a
full xserver/xfce environment installed) while the restore-mode images
were created specifically to provision targets while using a minimal
footprint (not even networking is supported in restore-mode etc.).

This commit adds a new safemode-restore-image to be used by nioldos to
do its work. The actual provisioning logic (disk format, rootfs
population, etc) is moved to meta-nilrt to not have the provisioning
logic split between layers.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>