safemode-image: always run do_install
The SAFEMODE_PAYLOAD_PATH variable doesn't change and so the safemode
contents are always reused from sstate accross builds. We need to
avoid this as each payload is different even though the variable is
identical, so instruct bitbake to always run do_install even though
nothing appears to be changed.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>